requirements.txt
wheel
Django==3.2.8
django-cors-headers==3.5.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
psycopg2-binary==2.9.1
PyJWT==1.7.1
python-decouple==3.5
requests==2.26.0
django-shopping-cart==0.1
django-phonenumber-field==5.0.0
phonenumbers==8.12.19
